created component to add new category
