--- client/src/components/AddCategory.jsx
@@ -1,5 +1,0 @@
-import { Fragment } from "react";
-
-export default function AddNewCategory() {
-	return <Fragment></Fragment>;
-}
